export tiles to files
